Dockerfile | Squashed the docker file for better caching
Added apt-get update on the same line as apt-get install. Having apt-get update on seperate line will actually make the docker image cache it. Therefore if one person wanted to append/add to the apt-get install the apt-get update will not be executed and therefore the packages wouldn't be updated.